refactor test_kickoff agents to test
