+++ server/test/test_kickoff.py
